pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "marketing-is-easy"
version = "0.2.0"
description = "Automated social media content generation and publishing for BioMaxing"
requires-python = ">=3.12"

dependencies = [
    # Core
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.6,<3",
    "python-dotenv>=1.0,<2",

    # Database
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",

    # Redis
    "redis[hiredis]>=5.0,<6",

    # Telegram Bot
    "aiogram>=3.15,<4",

    # AI / LLM
    "httpx>=0.28,<1",
    "openai>=1.60,<2",

    # Image
    "Pillow>=11.0,<12",

    # Video
    "moviepy>=2.2,<3",
    "pysrt>=1.1,<2",

    # YouTube
    "google-api-python-client>=2.150,<3",
    "google-auth>=2.36,<3",
    "google-auth-oauthlib>=1.2,<2",

    # Utilities
    "structlog>=24.0,<25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
    "ruff>=0.9",
    "mypy>=1.13",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
